initial

+++ application/models.py
@@ -101,7 +101,7 @@ class Base(Protected):
     @property
     def related_wall_instance(self):
         try:
-            return self.parent.first().container.wall
+            return self.document_set.first().container.wall
         except AttributeError:
             try:  # Container or Port
                 return self.wall
@@ -118,6 +118,13 @@ class Base(Protected):
     @property
     def version(self):
         return _to_hash(self.last_update.isoformat())
+
+    def copy(self):
+        clone = self.__class__.objects.get(pk=self.pk)
+        clone.id = None
+        clone.pk = None
+        clone.save()
+        return clone
 
 
 class SyncManager(Base):
@@ -173,9 +180,8 @@ class SyncManager(Base):
                     instance.save()
 
     @classmethod
-    def has_any_sync_widget(cls, user, sync_id=None):  # Empty sync_id is valid
-        # Validate sync_id. Find any object that belong to user and has id equal to requested sync_id
-        return not sync_id or cls.get_reachable(user).filter(pk=sync_id).exists()
+    def has_any_sync_widget(cls, sync_id):
+        return cls.objects.filter(pk=sync_id).exists()
 
 
 class Wall(SyncManager):
@@ -222,6 +228,13 @@ class Wall(SyncManager):
         container.save()
         return container
 
+    def copy(self):
+        clone = super().copy()
+        for container in self.container_set.all():
+            clone.container_set.add(container.copy())
+        clone.save()
+        return clone
+
 
 class Container(SyncManager):
     type = 'container'
@@ -249,6 +262,28 @@ class Container(SyncManager):
     def delete(self, *args, **kwargs):
         self.propagate_instance_deleted()
         super().delete(*args, **kwargs)
+
+    def copy(self):
+        clone = super().copy()
+        for w in self.simple_text_set.all():
+            clone.simple_text_set.add(w.copy())
+
+        for w in self.url_set.all():
+            clone.url_set.add(w.copy())
+
+        for w in self.simple_list_set.all():
+            clone.simple_list_set.add(w.copy())
+
+        for w in self.counter_set.all():
+            clone.counter_set.add(w.copy())
+
+        for w in self.simple_switch_set.all():
+            clone.simple_switch_set.add(w.copy())
+
+        for w in self.document_set.all():
+            clone.document_set.add(w.copy())
+        clone.save()
+        return clone
 
 
 class Port(SyncManager):
@@ -314,7 +349,7 @@ class Port(SyncManager):
 
 
 class Source(Base):
-    _wall_path = 'parent__container__wall__'
+    _wall_path = 'document_set__container__wall__'
     file = models.FileField(upload_to='%Y/%m/%d/', null=True)
 
     @classmethod
@@ -453,15 +488,16 @@ class Document(Widget):
 
     container = models.ForeignKey(Container, on_delete=models.CASCADE, related_name='document_set')
 
-    source = models.ForeignKey(Source, blank=True, on_delete=models.SET_NULL, null=True, related_name='parent')
+    # When copy widget - refer the same source (do not copy real file)
+    source = models.ForeignKey(Source, blank=True, on_delete=models.SET_NULL, null=True, related_name='document_set')
 
     def delete(self, *args, **kwargs):
-        if not self.has_any_sync_widget:
+        if not self.has_any_sync_widget(self.sync_id) and len(self.source.document_set.all()) == 1:
             self.source.delete_file()
-        try:
-            self.source.delete()
-        except AttributeError:
-            pass  # Source has been deleted
+            try:
+                self.source.delete()
+            except AttributeError:
+                pass  # Source has been deleted
         super().delete(*args, **kwargs)
